Standardize Slack Notifier We should not have `slack_` prefix
