+++ docs/apache-airflow-providers-slack/notifications/slack_notifier_howto_guide.rst
@@ -20,7 +20,7 @@ How-to Guide for Slack notifications
 
 Introduction
 ------------
-Slack notifier (:class:`airflow.providers.slack.notifications.slack_notifier.SlackNotifier`) allows users to send
+Slack notifier (:class:`airflow.providers.slack.notifications.slack.SlackNotifier`) allows users to send
 messages to a slack channel using the various ``on_*_callbacks`` at both the DAG level and Task level
 
 Example Code:
@@ -31,7 +31,7 @@ Example Code:
     from datetime import datetime
     from airflow import DAG
     from airflow.operators.bash import BashOperator
-    from airflow.providers.slack.notifications.slack_notifier import send_slack_notification
+    from airflow.providers.slack.notifications.slack import send_slack_notification
 
     with DAG(
         start_date=datetime(2023, 1, 1),
