+++ tests/providers/slack/notifications/test_slack.py
@@ -24,7 +24,7 @@ import pytest
 from airflow.models import Connection
 from airflow.operators.empty import EmptyOperator
 from airflow.providers.slack.hooks.slack import SlackHook
-from airflow.providers.slack.notifications.slack_notifier import SlackNotifier, send_slack_notification
+from airflow.providers.slack.notifications.slack import SlackNotifier, send_slack_notification
 
 SLACK_API_DEFAULT_CONN_ID = SlackHook.default_conn_name
 
@@ -47,7 +47,7 @@ def slack_api_connections():
 
 
 class TestSlackNotifier:
-    @mock.patch("airflow.providers.slack.notifications.slack_notifier.SlackHook")
+    @mock.patch("airflow.providers.slack.notifications.slack.SlackHook")
     def test_slack_notifier(self, mock_slack_hook, slack_api_connections, dag_maker):
         with dag_maker("test_slack_notifier") as dag:
             EmptyOperator(task_id="task1")
@@ -66,7 +66,7 @@ class TestSlackNotifier:
             },
         )
 
-    @mock.patch("airflow.providers.slack.notifications.slack_notifier.SlackHook")
+    @mock.patch("airflow.providers.slack.notifications.slack.SlackHook")
     def test_slack_notifier_with_notifier_class(self, mock_slack_hook, slack_api_connections, dag_maker):
         with dag_maker("test_slack_notifier") as dag:
             EmptyOperator(task_id="task1")
@@ -85,7 +85,7 @@ class TestSlackNotifier:
             },
         )
 
-    @mock.patch("airflow.providers.slack.notifications.slack_notifier.SlackHook")
+    @mock.patch("airflow.providers.slack.notifications.slack.SlackHook")
     def test_slack_notifier_templated(self, mock_slack_hook, slack_api_connections, dag_maker):
         with dag_maker("test_slack_notifier") as dag:
             EmptyOperator(task_id="task1")
